+++ packages/devtools_server/lib/src/handlers.dart
@@ -82,8 +82,6 @@ class ServerApi {
     switch (request.url.path) {
       case 'api/logScreenView':
         return api.logScreenView(request);
-      case 'api/getUserId':
-        return api.getUserId(request);
       default:
         return api.notImplemented(request);
     }
@@ -94,14 +92,6 @@ class ServerApi {
   /// In the open-source version of DevTools, Google Analytics handles this
   /// without any need to involve the server.
   FutureOr<shelf.Response> logScreenView(shelf.Request request) =>
-      notImplemented(request);
-
-  /// Gets a user's id for use in analytics and for distinguishing internal
-  /// users from external users.
-  ///
-  /// This endpoint is not supported externally and will only be implemented for
-  /// the version of DevTools used inside Google.
-  FutureOr<shelf.Response> getUserId(shelf.Request request) =>
       notImplemented(request);
 
   /// A [shelf.Response] for API calls that have not been implemented in this